requirements.txt
matplotlib
orjson
python-dotenv
requests
scikit-learn